pyproject.toml
[project]
name = "lua-kitchen-back"
version = "0.1.0"
description = ""
authors = [
    {name = "LimaXII_",email = "luccasdasilvalima@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "fastapi (>=0.129.2,<0.130.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "pydantic (>=2.12.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
